pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spendsmart"
version = "0.1.0"
description = "A personal budgeting copilot"
authors = [
    { "name" = "Nick Amaya" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
    "SQLAlchemy",
    "textual-dev"
]

[tool.hatch.build.targets.wheel]
packages = ["src/spendsmart"]